requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bug==0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-filter==22.1
idna==3.3
introcs==1.3.1
numpy==1.23.0
oauthlib==3.2.2
Pillow==9.1.1
pyaml==21.10.1
pycparser==2.21
PyJWT==2.6.0
pyTelegramBotAPI==4.7.0
python-dotenv==0.21.0
python3-openid==3.2.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
telebot==0.0.4
tzdata==2022.2
urllib3==1.26.9
zoom==0.0.15
